requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
deprecation==2.1.0
discord==2.3.2
discord-ext-slash==0.3.0
discord.py==2.5.0
Flask==3.1.0
frozenlist==1.5.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
multidict==6.1.0
packaging==24.2
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
psutil==5.9.8
six==1.17.0
sniffio==1.3.1
StrEnum==0.4.15
supabase==1.2.0
typing_extensions==4.12.2
Werkzeug==3.1.3
yarl==1.18.3
